# Handle timeout required by sdbus API