This PR adds correct handling of the timeout behavior mandated by the sdbus API. 

According to [the docs](https://www.freedesktop.org/software/systemd/man/latest/sd_bus_get_fd.html#), two conditions must be met:
- the file descriptor returned by sd_bus_get_fd() should be polled for the events indicated by sd_bus_get_events()
- and the I/O call should block for that up to the timeout returned by sd_bus_get_timeout()

However, since `loop.add_reader` doesn't support a timeout parameter, we create a separate timeout event with `loop.call_later`.

More precisely, depending on the return value of `sd_bus_get_timeout()`:
- if `UINT64_MAX`: no timeout is created
- if `0`: processing called immediately with `loop.call_soon`
- else: timeout created with `loop.call_later`