@ofacklam 2024-07-31
Hi @igo95862 !

Thanks for the quick reply and for the library in general, really handy!

Personally I didn't see any other way to integrate this with asyncio. If I understand correctly, the problem is not that you want to control the asyncio event loop's select policy, but rather you want to tell it to call **your** `sdbus_process` callback at the latest at that point in time.

However I would be glad to discuss other solutions as well :)

In our current project we are using the library with this patch since we were running into issues with the newer version 0.12.0. There are cases were both `sd_bus_get_events()` and `sd_bus_get_timeout()` return 0 to signal that data is already available and processing should be done immediately. In that case the current implementation deadlocks since no reader or writer will be registered anymore.