@igo95862 2024-07-31
I didn't know there was a deadlock which makes it a much bigger issue.

Looking at the sd-bus source code:

https://github.com/systemd/systemd/blob/941a12dcba57f6673230a9c413738c51374d2998/src/libsystemd/sd-bus/sd-bus.c#L2610

Looks like `sd_bus_get_events` returns POLLIN if its read queue is empty.

https://github.com/systemd/systemd/blob/941a12dcba57f6673230a9c413738c51374d2998/src/libsystemd/sd-bus/sd-bus.c#L2651

`sd_bus_get_timeout` returns zero when the read queue is not empty.

I wonder why calling the `sd_bus_process` repeatedly until it returns zero can leave some messages in the read queue?