@ofacklam 2024-07-31
Looking at the code again, it seems I did some synchronous calls on the same connection previously (a kind of app initialization logic) before starting the async calls. That might be the reason behind the strange behavior, and I guess it's not really an expected use case. So the issue is probably less severe than I thought.

Still I would argue that relying on an implementation detail of sdbus regarding timing behavior seems somewhat dangerous.