@igo95862 2024-08-02
> Looking at the code again, it seems I did some synchronous calls on the same connection previously (a kind of app initialization logic) before starting the async calls.

Is there some initialization that cannot be done with async code. Will the deadlock still happen if all code is properly async?

I will look in to this timeout integration this weekend. I have a few ideas how to make it more native.