@igo95862 2024-08-04
I looked in to the Linux APIs and asyncio and I have an interesting idea that I want to try.

I also want to setup benchmarks. One of my concerns is that constantly arming and disarming asyncio timer handles will have a negative effect on performance.