@igo95862 2024-08-11
I added initial benchmarks in 7e970a6f213c2ba1bba48f7b5b558eabe4f15747.

There definitely is performance impact of this change:

Before:

```
sdbus_async_ping: Mean +- std dev: 36.5 us +- 3.2 us
sdbus_async_ping_gather: Mean +- std dev: 14.0 us +- 1.0 us
```

After:

```
sdbus_async_ping: Mean +- std dev: 41.6 us +- 3.8 us
sdbus_async_ping_gather: Mean +- std dev: 17.3 us +- 0.4 us
```

My plan is to use `timerfd` because it seems to be reconfigurable meaning the timer can be set multiple times without having to create any new Python objects.