@ofacklam 2024-08-12
Your approach with `timerfd` sounds very promising. I'm looking forward to trying that out. Let me know if I can provide any assistance :)