@igo95862 2024-08-18
I created initial draft in 6bee5603e4ac862efa4aeda80661be5096412908. @ofacklam could you give it a test.

The performance impact seems to be much smaller:

```
.....................
sdbus_async_ping: Mean +- std dev: 37.0 us +- 2.3 us
.....................
sdbus_async_ping_gather: Mean +- std dev: 14.8 us +- 0.8 us
```

I also added a test there I set bus timeout value to 0.01 seconds and check that the `DbusNoReplyError` gets raised in time. Would be interesting adding a test that replicates your deadlock.