@ofacklam 2024-08-19
Hi @igo95862 

I cherry-picked your commit https://github.com/python-sdbus/python-sdbus/commit/6bee5603e4ac862efa4aeda80661be5096412908 onto the 0.12.0 release.

With this change, my "deadlock" seems to have disappeared, so it solves the problem just as well as my proposal 🎉 

Regarding normal timeouts of method calls, I can't forcibly reproduce it on my end (it only happens due to very rare hardware issues), however your unit test seems to cover that quite well.

Thanks for that fix, it looks really nice! Just one note: you rely on timerfd == 0 for checking whether it's initialized. Are you sure 0 can never be a valid timerfd value?