@igo95862 2024-08-19
> Just one note: you rely on timerfd == 0 for checking whether it's initialized. Are you sure 0 can never be a valid timerfd value?

Looks like if you close stdin file descriptor 0 then 0 can be reused for any other file descriptor so it definitely can be an issue.

I originally wanted to use -1 but I encountered issues reliably setting that in SdBus init method. I will revisit this before releasing final version.