@igo95862 2024-09-01
I added a `PyObject*` pointer to the bus struct which holds the timer fd Python integer in dcc61aa22e722d09a6e70aaa63bfca6fa37e3649. This way it can reliably tell if the timer was initialized by comparing it to NULL. The struct will have both Python and C integer representation so that it can be easily accessed from both.

I think I try to make more progress this week and release the 0.12.1 soon with this fix.