@igo95862 2024-09-22
Alright I merged the fe49edb17582d77c24e8ae9fd9698503db5e2ad1 that addresses the bus timeouts.