@igo95862 2024-10-30
I will probably switch to the `sd_event_get_fd` in the future version as it provides a single file descriptor specifically designed to integrate with existing event loops.

https://man.archlinux.org/man/sd_bus_attach_event.3.en

https://man.archlinux.org/man/sd_event_get_fd.3.en